pyproject.toml
[project]
name = "transformer-training-experimentation-pipeline"
version = "0.1.0"
description = "Scalable Transformer Training & Experimentation Pipeline with MLflow and Ray"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.8.4",
    "mlflow>=3.10.1",
    "psutil>=7.2.2",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "ray[default]>=2.54.1",
    "rich>=14.3.3",
    "scikit-learn>=1.8.0",
    "torch>=2.11.0",
    "transformers>=5.3.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "slow: marks tests that require model downloads or full training runs (deselect with '-m \"not slow\"')",
]
filterwarnings = [
    "ignore::FutureWarning:mlflow",
    "ignore::UserWarning:sklearn",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pipeline"]

# ---------------------------------------------------------------------------
# Ruff — linter + formatter
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes (unused imports, undefined names, etc.)
    "I",    # isort (import ordering)
    "UP",   # pyupgrade (modernise Python syntax)
    "B",    # flake8-bugbear (likely bugs and design issues)
    "SIM",  # flake8-simplify (simplifiable code patterns)
]
ignore = [
    "E501",   # line too long — ruff format handles line length
    "SIM108", # ternary operator — sometimes less readable than if/else
    "B008",   # do not perform function calls in default arguments
]
# Allow intentional noqa suppressions already in the codebase:
#   PLC0415 — import-outside-toplevel (used in Ray workers intentionally)
#   BLE001  — blind exception catch (used in Ray error handler intentionally)

[tool.ruff.lint.isort]
known-first-party = ["pipeline"]
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
